move refund transaction signing onto cfd
